Provide spdlog external library
The library enables a unified logging mechanism in Faunus. It is compiled
to a standalone object file libspdlog.a. When the preprocessor variable
SPDLOG_COMPILED_LIB is defined prior the header inclusion in Faunus, e.g.,
as a compiler argument, the object file is used instead of the inefficient
header-only mechanism. Issue #160.