Reduced header inter-dependencies
This is work in progess to reduce compilation times by reducing
header inter-dependencies and forward declarations. In particular
in this commit we have used forward declaration of SPD in the core.h
header which is included in almost all other files. Finally this
commit pushed tests in core.h and atomdata.h to separate test
files (note that core tests are currently disabled - will be fixed
shortly)