Add an aliasing of the mixing potentials coefficient names
To unify the user interface, the pair potentials coefficients are named
as sigma and eps in the topology. However if multiple potentials are
used at the same time, different value sets may be needed. Hence we
allow to create an alias, e.g., sigma: sigma_hs, which overrides the
eponymous default mapping. The naming of variables was adjusted
accordingly.

AtomData from_json function is now scavenging like a vulture, eating
anything looking like a number. If a builder for the state object was
introduced, the potentials could announce their parameter sets in
advance, enabling much cleaner way to get the topology parameters.

AtomData parameters, except sigma, are not access in any time-critical
part of the code, hence simple string-double map is used. A copy of
sigma is temporarily kept in an AtomData member.

Closes #171.