Remove PowerSASA
PowerSASA library is broken, unmaintained and undocumented. When used
to compute SASA of a hard sphere chain, exceptions are raised. They
are caused by limited precision of floating point numbers:
The coordinates of the same point are computed differently, depending
on the initial values used. Hence the point can be located at either
side of the boundary. This creates impossible states with odd number
of vertices in sets because the point at the boundary is bined
differently every time. The issue is completely unaddressed in the
code.