Allow a dot in ruamel.yaml (sub)module name
In the upstream, yaml is a submodule of ruamel. Hence the proper
import syntax includes a dot. However anaconda, unlike pip, packs
the submodule as a standalone module with an underscore instead of
the dot. Now we support both ways of importing the (sub)module,
with the pyyaml module option as a fallback.