Use composition over inheritance for Move::Propagator
Move::Propagator is not a vector itself but contains vector(s).

Furthermore, the BasePointerVector class is changed to follow closer
the std::vector interface. The original push_back method is renamed
to emplace_back and a new implementation of push_back is added, hence
already constructed instances can be appended.