Fail example tests if Faunus run fails
The output file shall be compared with its corresponding pattern only
if the Faunus run succeeded. That prevents false-positive results when
a correct output file has already existed and it has not been
overwritten, e.g., due to segfault or raised exception.

Note that the exit code of a semicolon-delimited sequence of commands
is the exit code of the last command, in contrast to the &&-delimited
sequence.

Also fix comparison within the relative tolerance in the tests.