Refactor and paralellise scattering analysis
OpenMP is enabled for loops in the sample methods that have O(N^2)
complexity, where N is number of particles.

Furthermore, the sample methods are rewritten to allow automatic
vectorisation of trigonometric functions in GCC by using libmvec
vector math library when --ffast-math is enabled. This speeds up
the analysis by factor of 4 on modern processors.