Enable Intel Threading Building Blocks (#325)
If ENABLE_TBB, Intel's threading building blocks
is downloaded and linked to all faunus targets. This
enables the use of parallel `std::execution` policies
currently available in GCC 9.1 and onwards. As exemplified
in `src/energy.h`, a macro can be used to deternine if
parallel execution is supported by the compiler.

The linked TBB is a clone, currently of TBB 2020.2 with
added CMake build, which enables simple download from our
build system.