Resurrect parallel tempering move
This fixes several problems with random number
states that would cause MPI deadlocks. Support
for fluctuating particles is temporarily disabled.

Test notebook added for running an 1D system with
and without tempering. Still needs to be added as
a test target.

Partial refactoring of ParallelTempering but still
plenty of things to improve in terms of safety and
clarity