Binary search for bond search in Bonded energy (#335)
Use binary search for bond energy calculation

When only a subset of particles are moved, the
previous algorithm would be of complexity O(N^2).
For large polymers this could be a significant bottle-neck.
Here we exploit that the list of changed particle index are
sorted, allowing for binary search. 10-100 times speedup have
been observed for large polymers.