Parallel Temper fix and example (#333)
This fixes several problems with random number
states that would cause MPI deadlocks. Support
for fluctuating particles is temporarily disabled.

Test notebook added for running an 1D system with
and without tempering. Still needs to be added as
a test target.

Partial refactoring of ParallelTempering but still
plenty of things to improve in terms of safety and
clarity.

* Add temperature scale to 2D example energy

* Resurrect parallel tempering move

* Add missing macro guard for MPI

* Update mass centers

* Add particle update function to Space

* Add unit test to Space::updateParticle()

* Adds `SpaceFactory::makeWater()` used in the unittest.

* Change order of volume/particle update

* Rearrange updateParticles test case

* Fix hexagonal prism failing test

* Add 'temper' to test suite w. statistical test