Make conformation loader respect `keepcharges`
The molecular topology keyword `keepcharges` can be used
to accept or reject charges in PQR files when using `structure`.
This was wrongly ignored when loading conformations using `traj` where
the file charges would always be used over the topology or `atomlist`
values. Related to conformational swapmoves combined with speciation
moves.