Add `copy_policy` to conformation swap move (#354)
`Move::ConformationSwap` now uses a variable copy policy describing how the conformation is copied in top if the target group in the simulation cell. For systems with atomic identify swaps, it can be useful to for example copy only positions in order not to alter the net charge.
In addition to this, we found an unexpected behavior when using conformation swap moves with speciation MC. Particles found now and then end up outside the simulation box and caught by `Analysis::Sanity` (which is partly re-written with better log output). The source of the problem was in the Metropolis acceptance policies where a cell collision would appear as a NaN value, which was wrongly accepted. Solved by always rejecting if trial energy is positive infinity.

* Add `copy_policy` flag

* Add to schema file

* Different policies are available: `all`, `positions`, `charges`

* Fix missing header in random.h

* Make conformation loader respect `keepcharges`

* Manual update

* Center conformations after loading from disk

* Rewrite Sanity check analysis with more logging if cell collision

* Check if conformation fit in PBC

* Add logging info and infinity reject check

* Add Group::getAtomIndex

* Make clang-format hook work with newer homebrew installations