Discrete angle integration (#427)
This adds the possibility to pre-calculate the interaction energy of two rigid bodies over a discrete angular space. Eventually this will be used to quickly obtain many-body interaction energies.

* Add angular scan move

* Add Geometry::TwobodyAngles

Class for iterating over the full angular space of
two rigid bodies. There is also a derived class with
an iterator-like state for stepping through angles.

* Add visual studio code extensions

* Add preface actions

* Add json schema for preface section

* Add (free) energy analysis

* Add anglescan example

Included Jupyter notebook for comparing
MC with discrete angle integration for a
dipolar system (CPPM)

* Update documentation